recoded with classes
