+++ scores.txt
@@ -1,0 +1,1 @@
+2.73, 4.713, 2.713, 2.712, 14.632, 2.712, 4.713, 5.592, 3.434, 5.43, 6.712, 2.71, 2.715, 3.444, 6.496, 5.444, 2.728, 5.433, 7.431, 5.453, 8.785, 5.428, 5.447, 5.43, 2.711, 2.721, 5.445, 5.429, 2.713, 12.697, 5.91, 28.714, 14.702, 6.812, 2.717, 2.712, 2.715, 10.734, 2.713, 2.725, 2.718, 2.717, 2.713, 2.716, 2.721, 4.719, 2.717, 2.712, 2.715, 2.714, 2.714, 2.714, 2.712, 2.728, 10.613, 14.737, 4.695, 2.723, 4.714, 2.714, 2.707, 2.713, 2.728, 2.712, 2.724, 4.724, 2.728, 2.716, 2.716, 2.72, 2.76, 2.72, 2.712, 2.784, 4.716, 2.715, 4.68, 0.694, 0.695, 8.645, 4.616, 2.712, 2.73, 4.777, 10.766, 4.714, 6.668, 2.714, 2.718, 2.717, 2.718, 2.716, 2.714, 2.718, 6.752, 2.716, 2.717, 2.714, 2.728, 2.718, 2.71, 2.745, 2.717, 2.714, 2.719, 2.714, 2.715, 8.679, 2.715, 2.715, 8.664, 4.712, 6.726, 28.711,
